Rename `infra.util` to `infra.config`
This is a more accurate name for what is in this module now.
