+++ pulumi/infra/service_queue.py
@@ -2,8 +2,7 @@ import json
 from typing import Optional
 
 import pulumi_aws as aws
-from infra import util
-from infra.util import AWS_ACCOUNT_ID, DEPLOYMENT_NAME
+from infra.config import AWS_ACCOUNT_ID, DEPLOYMENT_NAME, import_aware_opts
 
 import pulumi
 
@@ -47,7 +46,7 @@ class ServiceQueue(pulumi.ComponentResource):
             name=physical_dead_letter_name,
             message_retention_seconds=message_retention_seconds,
             visibility_timeout_seconds=30,
-            opts=util.import_aware_opts(
+            opts=import_aware_opts(
                 f"{queue_import_prefix}/{physical_dead_letter_name}",
                 parent=self,
                 delete_before_replace=True,
@@ -62,7 +61,7 @@ class ServiceQueue(pulumi.ComponentResource):
             message_retention_seconds=message_retention_seconds,
             visibility_timeout_seconds=360,
             redrive_policy=self.dead_letter_queue.arn.apply(redrive_policy),
-            opts=util.import_aware_opts(
+            opts=import_aware_opts(
                 f"{queue_import_prefix}/{physical_retry_name}",
                 parent=self,
                 delete_before_replace=True,
@@ -77,7 +76,7 @@ class ServiceQueue(pulumi.ComponentResource):
             message_retention_seconds=message_retention_seconds,
             visibility_timeout_seconds=180,
             redrive_policy=self.retry_queue.arn.apply(redrive_policy),
-            opts=util.import_aware_opts(
+            opts=import_aware_opts(
                 f"{queue_import_prefix}/{physical_queue_name}",
                 parent=self,
                 delete_before_replace=True,
